+++ requirements.txt
@@ -8,12 +8,9 @@ h11==0.14.0
 httpcore==1.0.4
 httpx==0.27.0
 idna==3.6
-nodeenv==1.8.0
 pydantic==2.6.3
 pydantic_core==2.16.3
-pyright==1.1.362
 selectolax==0.3.20
-setuptools==69.5.1
 sniffio==1.3.1
 starlette==0.36.3
 typing_extensions==4.10.0
